+++ TabletopTweaks/Bugfixes/Features/Feats.cs
@@ -1,10 +1,12 @@
 ﻿using HarmonyLib;
+using Kingmaker.Armies.TacticalCombat;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.Designers.EventConditionActionSystem.Actions;
 using Kingmaker.Designers.Mechanics.Facts;
+using Kingmaker.ElementsSystem;
 using Kingmaker.EntitySystem.Entities;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
@@ -19,7 +21,7 @@ using Kingmaker.UnitLogic.FactLogic;
 using Kingmaker.UnitLogic.Mechanics;
 using Kingmaker.UnitLogic.Mechanics.Actions;
 using Kingmaker.UnitLogic.Mechanics.Components;
-using Kingmaker.UnitLogic.Parts;
+using Kingmaker.UnitLogic.Mechanics.Conditions;
 using Kingmaker.Utility;
 using System;
 using System.Collections.Generic;
@@ -51,6 +53,7 @@ namespace TabletopTweaks.Bugfixes.Features {
                 PatchBolsteredMetamagic();
                 PatchEmpowerMetamagic();
                 PatchMaximizeMetamagic();
+                PatchShatterDefenses();
                 PatchSlashingGrace();
                 PatchSpiritedCharge();
                 PatchWeaponFinesse();
@@ -251,6 +254,68 @@ namespace TabletopTweaks.Bugfixes.Features {
                     };
                 }
             }
+            static void PatchShatterDefenses() {
+                if (ModSettings.Fixes.Feats.IsDisabled("ShatterDefenses")) { return; }
+
+                var ShatterDefenses = Resources.GetBlueprint<BlueprintFeature>("61a17ccbbb3d79445b0926347ec07577");
+                var ShatterDefensesBuff = Resources.GetModBlueprint<BlueprintBuff>("ShatterDefensesBuff");
+                var ShatterDefensesMythicFeat = Resources.GetModBlueprint<BlueprintFeature>("ShatterDefensesMythicFeat");
+                var ShatterDefensesMythicBuff = Resources.GetModBlueprint<BlueprintBuff>("ShatterDefensesMythicBuff");
+
+                ShatterDefenses.RemoveComponents<AddMechanicsFeature>();
+                ShatterDefenses.RemoveComponents<AddFacts>();
+                ShatterDefenses.AddComponent<ShatterDefensesInitiator>(c => {
+                    c.Action = Helpers.CreateActionList(
+                        new Conditional {
+                            ConditionsChecker = new ConditionsChecker {
+                                Conditions = new Condition[] {
+                                    new ContextConditionHasCondition() {
+                                        Conditions = new Kingmaker.UnitLogic.UnitCondition[]{
+                                            Kingmaker.UnitLogic.UnitCondition.Shaken,
+                                            Kingmaker.UnitLogic.UnitCondition.Frightened
+                                        }
+                                    }
+                                }
+                            },
+                            IfTrue = Helpers.CreateActionList(
+                                new Conditional {
+                                    ConditionsChecker = new ConditionsChecker {
+                                        Conditions = new Condition[] {
+                                            new ContextConditionCasterHasFact {
+                                                m_Fact = ShatterDefensesMythicFeat.ToReference<BlueprintUnitFactReference>(),
+                                            }
+                                        }
+                                    },
+                                    IfTrue = Helpers.CreateActionList(
+                                        Helpers.Create<ContextActionApplyBuff>(a => {
+                                            a.m_Buff = ShatterDefensesMythicBuff.ToReference<BlueprintBuffReference>();
+                                            a.DurationValue = new ContextDurationValue() {
+                                                m_IsExtendable = false,
+                                                Rate = DurationRate.Rounds,
+                                                DiceCountValue = 0,
+                                                BonusValue = 2
+                                            };
+                                        })
+                                    ),
+                                    IfFalse = Helpers.CreateActionList(
+                                        Helpers.Create<ContextActionApplyBuff>(a => {
+                                            a.m_Buff = ShatterDefensesBuff.ToReference<BlueprintBuffReference>();
+                                            a.DurationValue = new ContextDurationValue() {
+                                                m_IsExtendable = false,
+                                                Rate = DurationRate.Rounds,
+                                                DiceCountValue = 0,
+                                                BonusValue = 2
+                                            };
+                                        })
+                                    ),
+                                }
+                            ),
+                            IfFalse = Helpers.CreateActionList(),
+                        }
+                    );
+                });
+                Main.LogPatch("Patched", ShatterDefenses);
+            }
             static void PatchSpiritedCharge() {
                 if (ModSettings.Fixes.Feats.IsDisabled("SpiritedCharge")) { return; }
 
@@ -286,6 +351,22 @@ namespace TabletopTweaks.Bugfixes.Features {
                 Main.LogPatch("Patched", WeaponFinesse);
             }
         }
+
+        [HarmonyPatch(typeof(RuleCheckTargetFlatFooted), "OnTrigger", new Type[] { typeof(RulebookEventContext) })]
+        static class RuleCheckTargetFlatFooted_ShatterDefenses_Patch {
+            private static readonly BlueprintBuff ShatterDefensesBuff = Resources.GetModBlueprint<BlueprintBuff>("ShatterDefensesBuff");
+            private static readonly BlueprintBuff ShatterDefensesMythicBuff = Resources.GetModBlueprint<BlueprintBuff>("ShatterDefensesMythicBuff");
+
+            static void Postfix(RuleCheckTargetFlatFooted __instance, RulebookEventContext context) {
+                if (ModSettings.Fixes.Feats.IsDisabled("ShatterDefenses")) { return; }
+                if (TacticalCombatHelper.IsActive) { return; }
+                bool HasShatterFromCaster = __instance.Target.Buffs.Enumerable
+                    .Any(buff => buff.Blueprint == ShatterDefensesBuff && buff.Context.MaybeCaster == __instance.Initiator);
+                bool HasMythicShatter = __instance.Target.Buffs.HasFact(ShatterDefensesMythicBuff);
+                __instance.IsFlatFooted |= HasShatterFromCaster || HasMythicShatter;
+            }
+        }
+
         [HarmonyPatch(typeof(MetamagicHelper), "GetBolsteredAreaEffectUnits", new Type[] { typeof(TargetWrapper) })]
         static class MetamagicHelper_GetBolsteredAreaEffectUnits_Patch {
             static void Postfix(TargetWrapper origin, ref List<UnitEntityData> __result) {
@@ -331,7 +412,7 @@ namespace TabletopTweaks.Bugfixes.Features {
                 //Looking for the arguments that define the object creation because searching for the object creation itself is hard
                 for (int i = 0; i < codes.Count; i++) {
                     if (codes[i].opcode == OpCodes.Call && codes[i].Calls(AbilityCustomVitalStrike_get_RowdyFeature)) {
-                        if (codes[i+2].opcode == OpCodes.Newobj) {
+                        if (codes[i + 2].opcode == OpCodes.Newobj) {
                             return i + 2;
                         }
                     }