+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 1.3.5
+* shatter defenses now works like in tabletop
+* added mythic shatter defenses
+* extra feat mythic feat can now only be taken once
+* extra mythic ability mythic feat can now only be taken once
+* improvements to natual armor stacking rules
+
 ## Version 1.3.4a
 * fixed issue where weapon materials didn't apply correctly
 * fixed issues where activiatable abilities with resources spent incorrecty